change password checker to modular format
